Adjust Database Types

+++ Database/Source/renderonline_machines.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `machines`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `machines` (
-  `machine_id` bigint unsigned NOT NULL AUTO_INCREMENT,
+  `machine_id` tinyint unsigned NOT NULL AUTO_INCREMENT,
   `ip_address` varchar(45) NOT NULL,
   `type` varchar(20) NOT NULL,
   `subscriptions` text NOT NULL,
@@ -52,4 +52,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2024-08-22  0:01:47
+-- Dump completed on 2024-08-24 13:34:15
